+++ src/main.rs
@@ -6,6 +6,7 @@ use anyhow::{Error, anyhow};
 use chrono::Utc;
 use clap::Parser;
 use sentry_protos::taskbroker::v1::consumer_service_server::ConsumerServiceServer;
+use taskbroker::push::updater::{EagerUpdater, LazyUpdater, Updater};
 use taskbroker::worker::{Worker, WorkerClient, WorkerMap};
 use tokio::signal::unix::SignalKind;
 use tokio::task::JoinHandle;
@@ -25,16 +26,16 @@ use taskbroker::kafka::admin::create_missing_topics;
 use taskbroker::kafka::consumer::start_consumer;
 use taskbroker::kafka::deserialize::{self, DeserializeConfig};
 use taskbroker::kafka::os_stream_writer::{OsStream, OsStreamWriter};
-use taskbroker::logging;
 use taskbroker::metrics;
 use taskbroker::processing_strategy;
 use taskbroker::push::PushPool;
 use taskbroker::runtime_config::RuntimeConfigManager;
-use taskbroker::store::adapters::postgres::{PostgresStore, PostgresStoreConfig};
+use taskbroker::store::adapters::postgres::{self, PostgresStore, PostgresStoreConfig};
 use taskbroker::store::adapters::sqlite::{SqliteStore, SqliteStoreConfig};
 use taskbroker::store::traits::ActivationStore;
 use taskbroker::upkeep::upkeep;
 use taskbroker::{Args, get_version};
+use taskbroker::{Run, logging};
 use taskbroker::{SERVICE_NAME, flusher};
 
 async fn log_task_completion<T: AsRef<str>>(name: T, task: JoinHandle<Result<(), Error>>) {
@@ -64,11 +65,19 @@ async fn main() -> Result<(), Error> {
     logging::init(logging::LoggingConfig::from_config(&config));
     metrics::init(metrics::MetricsConfig::from_config(&config));
 
+    if args.run == Run::Migrations {
+        return config.database_adapter.migrate(&config).await;
+    }
+
     let store: Arc<dyn ActivationStore> = match config.database_adapter {
         DatabaseAdapter::Sqlite => Arc::new(
             SqliteStore::new(&config.db_path, SqliteStoreConfig::from_config(&config)).await?,
         ),
         DatabaseAdapter::Postgres => {
+            if config.run_migrations {
+                postgres::migrate(&config).await?;
+            }
+
             Arc::new(PostgresStore::new(PostgresStoreConfig::from_config(&config)).await?)
         }
     };
@@ -285,7 +294,7 @@ async fn main() -> Result<(), Error> {
     let (sender, receiver) = flume::bounded(config.push_queue_size);
 
     // Initialize push and fetch pools
-    let push_pool = PushPool::new(receiver, config.clone(), store.clone());
+    let push_pool = PushPool::new(receiver, config.clone());
     let fetch_pool = FetchPool::new(sender, store.clone(), config.clone());
 
     // Initialize push threads
@@ -315,8 +324,17 @@ async fn main() -> Result<(), Error> {
             workers.push(map);
         }
 
+        // Create the correct kind of push updater
+        let updater = if config.batch_push_updates {
+            let lazy = LazyUpdater::new(config.clone(), store.clone());
+            Arc::new(lazy) as Arc<dyn Updater>
+        } else {
+            let eager = EagerUpdater::new(store.clone());
+            Arc::new(eager) as Arc<dyn Updater>
+        };
+
         Some(taskbroker::tokio::spawn(async move {
-            push_pool.start(workers).await
+            push_pool.start(workers, updater, store.clone()).await
         }))
     } else {
         None